chore(doc): document the target member of the Request struct
The target member wasn't as clear without documentation as it should
have been in the beginning.
People developing websites hev the habits to refer to routes but target
is way less often mentioned whereas it is used in the HTTP RFC.
So the target not only refers to the route but also include the query
parameters of the request.
So thew target is broader than the route.