fix: Request struct member order
In order to be able to use the brace initializer constructor of the
structure, we want the variables to be in the same order as they are
parsed.
So here an HTTP request contains the target before the version of HTTP
used so the members are now following this order.