fix: `fields` to `headers` in both the struct and the doc
As the modification of the member `fields` in the structure `Request`
was aimed, it makes sense to also change this in the examples and the
documentation.
It also made sense to change it in the `Response` structure since it
makes reference to the exact same thing whether it's in an Http request
or response.

BREAKING CHANGE: fields is now named headers